// src/app/Model/user.model.ts
export interface User{
    id: string;
    name: string;
    username: string;
    password: string;
    hasAccess:boolean

    // constructor(id:number,name:string,username:string,password:string){
    //     this.id=id;
    //     this.name=name;
    //     this.username=username;
    //     this.password=password
    // }

}